refactor form for settings page
